Move test classes to claw.tatsu.xcodeml.xnode.fortran
