+++ cx2t/unittest/claw/tatsu/xcodeml/xnode/fortran/FfunctionDefinitionTest.java
@@ -2,12 +2,12 @@
  * This file is released under terms of BSD license
  * See LICENSE file for more information
  */
-package claw.tatsu.xcodeml.xnode.common;
+package claw.tatsu.xcodeml.xnode.fortran;
 
 import claw.tatsu.common.CompilerDirective;
 import claw.tatsu.common.Context;
 import claw.tatsu.common.Target;
-import claw.tatsu.xcodeml.xnode.fortran.FfunctionDefinition;
+import claw.tatsu.xcodeml.xnode.common.XcodeProgram;
 import helper.TestConstant;
 import helper.XmlHelper;
 import org.junit.Test;
