+++ cx2t/unittest/claw/tatsu/xcodeml/xnode/fortran/FmoduleDefinitionTest.java
@@ -2,9 +2,11 @@
  * This file is released under terms of BSD license
  * See LICENSE file for more information
  */
-package claw.tatsu.xcodeml.xnode.common;
+package claw.tatsu.xcodeml.xnode.fortran;
 
-import claw.tatsu.xcodeml.xnode.fortran.FmoduleDefinition;
+import claw.tatsu.xcodeml.xnode.common.Xcode;
+import claw.tatsu.xcodeml.xnode.common.XcodeProgram;
+import claw.tatsu.xcodeml.xnode.common.Xnode;
 import helper.TestConstant;
 import helper.XmlHelper;
 import org.junit.Test;
