+++ cx2t/unittest/claw/tatsu/xcodeml/xnode/fortran/FortranTypeTest.java
@@ -2,10 +2,9 @@
  * This file is released under terms of BSD license
  * See LICENSE file for more information
  */
-package claw.tatsu.xcodeml.xnode.common;
+package claw.tatsu.xcodeml.xnode.fortran;
 
 import claw.tatsu.xcodeml.xnode.Xname;
-import claw.tatsu.xcodeml.xnode.fortran.FortranType;
 import org.junit.Test;
 
 import static org.junit.Assert.assertFalse;
